*: cleanup code
Re-order function definitions for better readability and reduce
some nested mess. This commit also fixes a bug in a CLI, where
passing an infinite file would hang the CLI as it tries to read
the whole file for a segment instead of just what is needed.
The new version also avoids an extra allocation in that same
code.